pyproject.toml
[project]
name = "assignment-5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0.0",
    "pillow>=9.0.0",
    "pytest>=8.4.2",
]
